create getMarkerValues utility function and implement it
